Move decorator test to separate file
